+++ src/App.jsx
@@ -1,6 +1,9 @@
 import HomepageIndex from "./pages/homepage/index"
 import GameLoader from "./components/loaders/game-loader"
 
+//Temporary
+import HomePage from "./pages/references"
+
 function App() {
   return (
     <>